Move all reply message reaction to `replayMessageReaction()` function
